+++ include/countly/storage_module_memory.hpp
@@ -19,12 +19,12 @@ public:
   ~StorageModuleMemory();
 
   void init() override;
-  int RQCount() override;
+  long long RQCount() override;
   void RQClearAll() override;
   virtual void RQRemoveFront() override;
-  const std::shared_ptr<DataEntry>RQPeekFront() override;
+  const std::shared_ptr<DataEntry> RQPeekFront() override;
   std::vector<std::shared_ptr<DataEntry>> RQPeekAll() override;
-  void RQRemoveFront(std::shared_ptr<DataEntry>request) override;
+  void RQRemoveFront(std::shared_ptr<DataEntry> request) override;
   void RQInsertAtEnd(const std::string &request) override;
 };
 } // namespace cly
